move input sanitization from JailConfig to BaseConfig
By making the sanitization part of the BaseConfig, also user configured defaults are sanitized.
